Move connect to net module
